+++ src/client/modules/AccountBook/AccountBookHistory/components/index.ts
@@ -1,0 +1,2 @@
+export { default as HistoryEmptyView } from './HistoryEmptyView';
+export { default as HistoryList } from './HistoryList';
